Overhaul v11 (changes in description of v10 as well)
V10:
- The data in the MyStats GUI will now autoupdate on data change.
- Got rid of unnecessary exception checks throughout the code.
- Implemented MySQL for the whole plugin.
- Error messages will no longer be supressed by async methods.
- Messages of the AH commands are now more in line with all other commands.
- Async methods are now actually async resulting in much less lag, they were accidentally put in sync.
- MyStats has been removed from the ah tab completion suggestions.
- Configurations are now cached on startup resulting in much less lag on specific events.
- Removed the prefix option for MySQL.
- Fixed a NPE in deathban message when the killer didn't have a type.
- Bans (server and player) are now saved in a navigatable hashmap for easier use.
- IP has been changed to lastknownip to be inline with lastknownname.
- Fixed a bug where commands and lightning were throwing an error because player left the game already.
- Fixed bug where combat tagging wasn't working as intended.
- Fixed bug where not having health (death) would throw an exception when joining the server.
- Altered the naming convention of a lot of classes to be inline with eachother.
- Life parts on playtime will now stop counting down when max life parts is achieved.
- Altered the cooldown variable types to be inline with eachother and Minecraft.
- Fixed a bug in the permission system.
- Removed the auto update checker for now since it was causing some unnecessary lag.
- Fixed the GUI fill method.
- Fixed a bug in the opening of the MyStats GUI where the GUI thought you were somebody else.
- The revive GUI has been made, slightly more performant.
- Got rid of all the bukkit runnables and changed them with completablefutures.
- The plugin now also keeps track of players banned through the plugin outside of the plugin.
- Fixed a lot of possible NPE's in the config deserialization (most of them weren't able to actually throw them, but just in case).
- Messages telling time will no longer show day/hour/minute/second when the time is 0, instead it will show days/hours/minutes/seconds.
- Fixed some hardcoded plugin names.
- Added the option to ban players permanently on death ban.
- Fixed some colour issues in the ban messages on Spigot.

v11:
- Added the permissions augmentedhardcore.bypass.ban.spectator. The permission will allow players to be banned into spectator made instead of off the server. After the ban runs out the player is able to execute the command /deathbanreset to start playing again.
- The revive feature now doesn't require the lives features to be enabled anymore.
- Fixed a NPE when removing a ban from the server that didn't exist.
- Playtime countdowns will now save every 5 minutes instead of every 1 minute.

+++ src/com/backtobedrock/augmentedhardcore/domain/data/ServerData.java
@@ -79,7 +79,9 @@ public class ServerData {
 
     public void removeBan(OfflinePlayer player) {
         Pair<Integer, Ban> banPair = this.ongoingBans.remove(player.getUniqueId());
-        this.plugin.getServerRepository().removeBanFromServerData(player, banPair);
+        if (banPair != null) {
+            this.plugin.getServerRepository().removeBanFromServerData(player, banPair);
+        }
     }
 
     public boolean isDeathBanned(OfflinePlayer player) {
